MeshViewer: add controls
