+++ gui/widgets/mesh_viewer/__init__.py
@@ -1,5 +1,7 @@
+from .render_widget import MeshRenderWidget
 from .viewer_widget import MeshViewer
 
 __all__ = [
-    "MeshViewer"
+    "MeshRenderWidget",
+    "MeshViewer",
 ]
